dmd 0.176

+++ src/attrib.c
@@ -743,7 +743,7 @@ void PragmaDeclaration::semantic(Scope *sc)
 		if (e->op == TOKstring)
 		{
 		    StringExp *se = (StringExp *)e;
-		    fprintf(stdmsg, "%.*s", se->len, se->string);
+		    fprintf(stdmsg, "%.*s", (int)se->len, se->string);
 		}
 		else
 		    error("string expected for message, not '%s'", e->toChars());
